+++ Mage/src/main/java/mage/abilities/dynamicvalue/common/GetXLoyaltyValue.java
@@ -14,7 +14,7 @@ public enum GetXLoyaltyValue implements DynamicValue {
 
     @Override
     public int calculate(Game game, Ability sourceAbility, Effect effect) {
-        return (int) CardUtil.getSourceCostsTag(game, sourceAbility, "X", 0);
+        return CardUtil.getSourceCostsTag(game, sourceAbility, "X", 0);
     }
 
     @Override
